Deleted original About component,  Renamed AboutGrid -> About
