--- src/components/Sections/About/AboutGrid.scss
@@ -1,191 +1,0 @@
-$breakpoint-splus: 600px;
-$breakpoint-mplus: 650px;
-$breakpoint-lplus: 800px;
-$breakpoint-xlplus: 1000px;
-$color-highlight-purple: #A259FF;
-$color-primary-purple: #6C63FF;
-$color-quaternary-purple: #6661c1;
-$color-primary-lightblue: #83D4ED;
-$color-dark-bg: #1B1B2D;
-
-@mixin min-media($width) {
-    @media only screen and (min-width: $width) {
-      @content;
-    }
-}
-
-@mixin max-media($width) {
-    @media only screen and (max-width: $width) {
-      @content;
-    }
-}
-
-.about {
-    padding: 25rem 2rem 30rem;
-    margin: 0 auto;
-    color: inherit;
-    position: relative;
-   
-    .about__header {
-        display: grid;
-        grid-template-rows: 0.1fr repeat(2, 0.4fr);
-        grid-template-columns: 0.2fr 1fr;
-        gap: 1.5rem; // remove this on desktop
-        font-size: 1.6rem;
-        padding: 3rem 0;
-
-        @include min-media($breakpoint-lplus) {
-            max-height: 960px;
-            // max-width: 1350px;
-        }
-
-        @include min-media($breakpoint-splus) {
-            grid-template-rows: 1fr 0.2fr 1fr;
-            grid-template-columns: repeat(2, 1fr);
-            place-items: center;
-            margin: 0 auto;
-            padding: 3rem 0.5rem;
-            font-size: 1.5rem;
-        }
-    }
-
-    .about__heading {
-        z-index: 2;
-        font-size: 4rem;
-        grid-column: 1 / span 2;
-    
-        @include min-media($breakpoint-splus){
-            // font-size: clamp( 4.5rem, 10vw, 150px);
-            font-size: 8.5vw;
-            margin: 0;
-        }
-    
-        // @include min-media($breakpoint-lplus) { 
-        //     font-size: 6rem;
-        // }
-
-        // @include min-media($breakpoint-xlplus) { 
-        //     font-size: 13rem;
-        // }
-
-        &:nth-of-type(1) {
-            grid-row: 1 / span 1; 
-
-            @include min-media($breakpoint-splus) {
-                grid-row: 1 / span 1; 
-                grid-column: 1 / span 1;
-            }
-        }
-
-        &:nth-of-type(2) {  
-            grid-row: 2 / span 1; 
-            @include min-media($breakpoint-splus) {
-                grid-column: 1 / span 2; 
-            }
-        }
-
-        &:nth-of-type(3) {
-            grid-row: 3 / span 1;
-            margin-bottom: 3rem;
-            color: $color-highlight-purple;
-
-            @include min-media($breakpoint-splus) {
-                grid-row: 3 / span 1;
-                grid-column: 2 / span 1;
-                margin-bottom: 0;
-            }
-        }
-    }
-
-    .about__text-container { 
-        position: relative;
-        overflow: hidden;
-        z-index: 2;
-        line-height: 3rem;
-        color: #dedeeb;
-        // width: 85%;
-        
-
-        @include min-media($breakpoint-lplus) {
-            font-size: 1.9rem; 
-        }
-
-        @include min-media($breakpoint-splus) {
-            line-height: 2.6rem;
-        }
-
-        
-        &:nth-of-type(1) {
-            grid-row: 4 / span 1; 
-            grid-column: 2 / span 1;
-            margin-top: 5rem;
-
-            @include min-media($breakpoint-splus) {
-                grid-row: 1 / span 1; 
-                // align-self: flex-end;
-                // justify-self: flex-start;
-            }
-
-            @include min-media($breakpoint-lplus) {
-                width: 70%;
-            }
-        }
-
-        &:nth-of-type(2) {
-            grid-row: 5 / span 1; 
-            grid-column: 2 / span 1;
-
-            @include min-media($breakpoint-splus) {
-                grid-column: 1 / span 1;
-                grid-row: 3 / span 1;
-                // justify-self: flex-start;
-            }
-
-            @include min-media($breakpoint-lplus) {
-                width: 70%;
-            }
-        };
-
-
-        &:nth-of-type(3) {
-            @include min-media($breakpoint-splus) {
-                align-self: flex-end;
-            }
-        }
-
-
-        p {
-            margin-bottom: 2rem;
-            font-size: 1.5rem;
-            color: $color-primary-purple;
-        }
-    }
-    
-    .img1 {
-        position: absolute;
-        height: 260px;
-        width: 200px;
-        z-index: 1;
-        bottom: -10%;
-        left: 2.5rem;
-
-        @include min-media($breakpoint-splus) {
-            height: 320px;
-            width: 250px;
-            left: 30%;
-            top: 80%;
-        }
-    }
-}
-
-.split-parent {
-  overflow: hidden;
-}
-
-.split-child {
-  display: inline-block;
-}
-
-.lines {
-    width: 50%;
-}
